refactor: consolidate related functionality into a single file
